Update access
closes #53
closes #52
